Fix 'actions updated' broken by 2487.
The idea was that dependents of NOTFILE targets (like "all")
shouldn't be T_FATE_NEWER (newer than their parents), but instead
left as T_FATE_STABLE.  That way the new 'jam -dc' (display
"causes") option didn't report spurious "newer" targets.  NOTFILE
targets have a 0 timestamp, making their children always look
newer, and the change was to suppress this.

But instead of checking for NOTFILE parents, the code checked
for missing parents.  And that broke 'actions updated', because
it relied on T_FATE_NEWER even if the parent was missing.

Now it checks for p->binding != T_BIND_UNBOUND (the mark of a
NOTFILE target), rather than p->binding == T_BIND_EXISTS (the
mark of an existing real target).

=== computer:1666: Change 39756 by seiwald@play-seiwald on 2003/01/03 14:53:11

[git-p4: depot-paths = "//guest/perforce_software/jam/src/": change = 2613]