Two Jambase changes: one fix, one enhancement.
The fix: if a Jamrules invoked SubDir to establish other roots,
and that Jamrules isn't in the current directory, the roots it
established were wrong.

The enhancement:  Undocumented support for SUBDIRRULES,
user-provided rules to invoke at the end of the SubDir rule,
and SUBDIRRESET, SUBDIR variables to reset (like SUBDIRC++FLAGS,
SUBDIRHDRS, etc) for each new SubDir.

Bug fix and change documented only in RELNOTES.

[git-p4: depot-paths = "//guest/perforce_software/jam/src/": change = 3070]