jam and 2004.2 p4 client for AS/400. Built on AS/400 V5R2 with native ILE C/C++ compiler. Note, this build reports itself as P4/AS400/2004.2/76944 but this change contains porting changes that were required to make the build work so the changelist number is incorrect. This can be resolved with a clean build for 2005.1 at a later date.
p4.sav and jam.sav are AS/400 "Save Files" that can be used to install
the software on an AS/400 machine. The p4.sav file includes the p4.cmd
and p4sync.cmd files we distributed with our older AS/400 builds. These
files define a forms-type interface for supplying the parameters to the
commands when you run them.

Porting change only. No functional change.


p4transfer.py: Transferred from production

[git-p4: depot-paths = "//guest/perforce_software/jam/src/": change = 9912]